@@ -1,1 +1,1 @@
-ruby-1.9.2-p320
+ruby-2.2.3
