@@ -1,213 +1,57 @@
-set :default_stage, 'staging'
-require 'capistrano/ext/multistage'
-## Generated with 'brightbox' on 2012-05-08 09:53:55 +0100
-gem 'brightbox', '>=2.3.9'
-require 'brightbox/recipes'
-require 'brightbox/passenger'
-require 'sidekiq/capistrano'
-set :sidekiq_processes, 1
+# config valid only for current version of Capistrano
+lock '3.4.0'
 
-set :generate_webserver_config, false
+set :application, 'magpie'
+set :repo_url, 'git@github.com:unepwcmc/magpie.git'
 
-ssh_options[:forward_agent] = true
+# Default branch is :master
+# ask :branch, `git rev-parse --abbrev-ref HEAD`.chomp
 
-set :rake, 'bundle exec rake'
-set :branch, fetch(:branch, "master")
 
-# The name of your application.  Used for deployment directory and filenames
-# and Apache configs. Should be unique on the Brightbox
-set :application, "magpie"
+set :branch, 'migration'
 
-# Target directory for the application on the web and app servers.
-set(:deploy_to) { File.join("", "home", user, application) }
+set :deploy_user, 'wcmc'
 
-# URL of your source repository. By default this will just upload
-# the local directory.  You should probably change this if you use
-# another repository, like git or subversion.
 
-set :repository,  "git@github.com:unepwcmc/magpie.git"
+# Default deploy_to directory is /var/www/my_app_name
+set :deploy_to, "/home/#{fetch(:deploy_user)}/#{fetch(:application)}"
+
+# Default value for :scm is :git
 set :scm, :git
 set :scm_username, "unepwcmc-read"
-set :deploy_via, :remote_cache
 
-### Other options you can set ##
-# Comma separated list of additional domains for Apache
-# set :domain_aliases, "www.example.com,dev.example.com"
 
-## Dependencies
-# Set the commands and gems that your application requires. e.g.
-# depend :remote, :gem, "will_paginate", ">=2.2.2"
-# depend :remote, :command, "brightbox"
-#
-# If you're using Bundler, then you don't need to specify your
-# gems here as well as there (and the bundler gem is installed for
-# you automatically)
-# 
-# Gem with a source (such as github)
-# depend :remote, :gem, "tmm1-amqp", ">=0.6.0", :source => "http://gems.github.com"
-# 
-# Specify your specific Rails version if it is not vendored
-# depend :remote, :gem, "rails", "=2.2.2"
-#
-# Set the apt packages your application or gems require. e.g.
-# depend :remote, :apt, "libxml2-dev"
+set :rvm_type, :user
+set :rvm_ruby_version, '2.2.3'
 
-## Local Shared Area
-# These are the list of files and directories that you want
-# to share between the releases of your application on a particular
-# server. It uses the same shared area as the log files.
-#
-# NOTE: local areas trump global areas, allowing you to have some
-# servers using local assets if required.
-#
-# So if you have an 'upload' directory in public, add 'public/upload'
-# to the :local_shared_dirs array.
-# If you want to share the database.yml add 'config/database.yml'
-# to the :local_shared_files array.
-#
-# The shared area is prepared with 'deploy:setup' and all the shared
-# items are symlinked in when the code is updated.
-# set :local_shared_dirs, %w(public/upload)
-set :local_shared_files, %w(config/database.yml config/cartodb_config.yml)
 
-## Global Shared Area
-# These are the list of files and directories that you want
-# to share between all releases of your application across all servers.
-# For it to work you need a directory on a network file server shared
-# between all your servers. Specify the path to the root of that area
-# in :global_shared_path. Defaults to the same value as :shared_path.
-# set :global_shared_path, "/srv/share/myapp"
-#
-# NOTE: local areas trump global areas, allowing you to have some
-# servers using local assets if required.
-#
-# Beyond that it is the same as the local shared area.
-# So if you have an 'upload' directory in public, add 'public/upload'
-# to the :global_shared_dirs array.
-# If you want to share the database.yml add 'config/database.yml'
-# to the :global_shared_files array.
-#
-# The shared area is prepared with 'deploy:setup' and all the shared
-# items are symlinked in when the code is updated.
-# set :global_shared_dirs, %w(public/upload)
-# set :global_shared_files, %w(config/database.yml)
 
-# SSL Certificates. If you specify an SSL certificate name then
-# the gem will create an 'https' configuration for this application
-# TODO: Upload and install the keys on the server
-# set :ssl_certificate, "/path/to/certificate/for/my_app.crt"
-# set :ssl_key, "/path/to/key/for/my_app.key
-# or
-# set :ssl_certificate, "name_of_installed_certificate"
+set :ssh_options, {
+  forward_agent: true,
+}
 
-## Static asset caching.
-# By default static assets served directly by the web server are
-# cached by the client web browser for 10 years, and cache invalidation
-# of static assets is handled by the Rails helpers using asset
-# timestamping.
-# You may need to adjust this value if you have hard coded static
-# assets, or other special cache requirements. The value is in seconds.
-# set :max_age, 315360000
 
-# SSH options. The forward agent option is used so that loopback logins
-# with keys work properly
-# ssh_options[:forward_agent] = true
+# Default value for :format is :pretty
+# set :format, :pretty
 
-# Forces a Pty so that svn+ssh repository access will work. You
-# don't need this if you are using a different SCM system. Note that
-# ptys stop shell startup scripts from running.
-default_run_options[:pty] = true
+# Default value for :log_level is :debug
+#set :log_level, :debug
 
-## Logrotation
-# Where the logs are stored. Defaults to <shared_path>/log
-# set :log_dir, "central/log/path"
-# The size at which to rotate a log. e.g 1G, 100M, 5M. Defaults to 100M
-# set :log_max_size, "100M"
-# How many old compressed logs to keep. Defaults to 10
-# set :log_keep, "10"
+# Default value for :pty is false
+set :pty, true
 
-## Version Control System
-# Which version control system. Defaults to subversion if there is
-# no 'set :scm' command.
-# set :scm, :git
-# set :scm_username, "rails"
-# set :scm_password, "mysecret"
-# or be explicit
-# set :scm, :subversion
+# Default value for :linked_files is []
 
-## Deployment settings
-# The brightbox gem deploys as the user 'rails' by default and
-# into the 'production' environment. You can change these as required.
-# set :user, "rails"
-# set :rails_env, :production
+set :linked_files, %w{config/database.yml config/cartodb_config.yml .env}
 
-## Command running settings
-# use_sudo is switched off by default so that commands are run
-# directly as 'user' by the run command. If you switch on sudo
-# make sure you set the :runner variable - which is the user the
-# capistrano default tasks use to execute commands.
-# NB. This just affects the default recipes unless you use the
-# 'try_sudo' command to run your commands.
-# set :use_sudo, false
-# set :runner, user## Passenger Configuration
-# Set the method of restarting passenger
-# Defaults to :hard which is used to instantly free up database connections
-# :soft uses the standard touch tmp/restart.txt which leaves database connections
-# lingering until the workers time out
-# set :passenger_restart_strategy, :hard
+# Default value for linked_dirs is []
+set :linked_dirs, fetch(:linked_dirs, []).push('log', 'tmp/pids', 'tmp/cache', 'tmp/sockets', 'vendor/bundle', 'public/system')
 
-task :setup_production_database_configuration do
-  the_host = Capistrano::CLI.ui.ask("Database IP address: ")
-  database_name = Capistrano::CLI.ui.ask("Database name: ")
-  database_user = Capistrano::CLI.ui.ask("Database username: ")
-  pg_password = Capistrano::CLI.password_prompt("Database user password: ")
 
-  require 'yaml'
+# Default value for default_env is {}
+# set :default_env, { path: "/opt/ruby/bin:$PATH" }
 
-  spec = {
-    "#{rails_env}" => {
-      "adapter" => "postgresql",
-      "database" => database_name,
-      "username" => database_user,
-      "host" => the_host,
-      "password" => pg_password
-    }
-  }
+# Default value for keep_releases is 5
+set :keep_releases, 5
 
-  run "mkdir -p #{shared_path}/config"
-  put(spec.to_yaml, "#{shared_path}/config/database.yml")
-end
-after "deploy:setup", :setup_production_database_configuration
-
-# CartoDB
-
-namespace :cartodb do
-  desc "Generate a cartodb configuration file"
-  task :build_configuration do
-    host = Capistrano::CLI.ui.ask("CartoDB host: ")
-    oauth_key = Capistrano::CLI.ui.ask("CartoDB key: ")
-    oauth_secret = Capistrano::CLI.ui.ask("CartoDB secret: ")
-    username = Capistrano::CLI.ui.ask("CartoDB username: ")
-    password = Capistrano::CLI.password_prompt("CartoDB password: ")
-
-    cartodb_options = {
-      "host" => host,
-      "oauth_key" => oauth_key,
-      "oauth_secret" => oauth_secret,
-      "username" => username,
-      "password" => password
-    }
-
-    config_options = {"production" => cartodb_options}.to_yaml
-    run "mkdir -p #{shared_path}/config"
-    put config_options, "#{shared_path}/config/cartodb_config.yml"
-  end
-
-  desc "Links the configuration file"
-  task :link_configuration_file do
-    run "ln -nsf #{shared_path}/config/cartodb_config.yml #{latest_release}/config/cartodb_config.yml"
-  end
-end
-
-after "deploy:setup", "cartodb:build_configuration"
-after "deploy:finalize_update", "cartodb:link_configuration_file"
+set :passenger_restart_with_touch, false
