@@ -1,5 +1,5 @@
-require 'simplecov'
-SimpleCov.start
+#require "codeclimate-test-reporter"
+#CodeClimate::TestReporter.start
 
 # This file is copied to spec/ when you run 'rails generate rspec:install'
 ENV["RAILS_ENV"] ||= 'test'
